requirements.txt
grpcio==1.69.0
protobuf==5.29.3
tabulate==0.9.0
